refactor package name; add logging and status tracking
